+++ composer.json
@@ -27,7 +27,7 @@
         "symfony/form":             "^2.7|^3.0",
         "symfony/framework-bundle": "^2.7|^3.0",
         "symfony/http-kernel":      "^2.7|^3.0",
-        "symfony/phpunit-bridge":   "^2.8|^3.0",
+        "symfony/phpunit-bridge":   "^2.7|^3.0",
         "symfony/translation":      "^2.7|^3.0",
         "symfony/validator":        "^2.7|^3.0"
     },
