+++ test/FormHandlerBundleTest.php
@@ -1,36 +1,30 @@
 <?php
-/**
- * @copyright 2017 Hostnet B.V.
- */
 namespace Hostnet\Bundle\FormHandlerBundle;
-
-use Hostnet\Bundle\FormHandlerBundle\DependencyInjection\Compiler\FormParamConverterCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 
 /**
- * @author Iltar van der Berg <ivanderberg@hostnet.nl>
- * @covers \Hostnet\Bundle\FormHandlerBundle\FormHandlerBundle
+ * @author Piotr Rzeczkowski <piotr@rzeka.net>
+ * @coversDefaultClass Hostnet\Bundle\FormHandlerBundle\FormHandlerBundle
  */
 class FormHandlerBundleTest extends \PHPUnit_Framework_TestCase
 {
-    public function testBuild()
+    public function testIsInstanceOfFormHandlerBundle()
+    {
+        $bundle = new FormHandlerBundle();
+
+        $this->assertTrue($bundle instanceof HostnetFormHandlerBundle);
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation The %s is deprecated. Use %s instead.
+     */
+    public function testIsDeprecated()
     {
         $container = new ContainerBuilder();
-
         $bundle = new FormHandlerBundle();
+
         $bundle->build($container);
 
-        $found = false;
-
-        foreach ($container->getCompilerPassConfig()->getBeforeOptimizationPasses() as $pass) {
-            if (!$pass instanceof FormParamConverterCompilerPass) {
-                continue;
-            }
-
-            $found = true;
-            break;
-        }
-
-        self::assertTrue($found, 'Expected to find a compiler pass instance of the FormParamConverterCompilerPass.');
     }
 }
